refactor(add comment): introduce hooks for this component
